// src/types.ts
export type DownsParameterKey =
  | 'facialAngle'
  | 'angleConvexity'
  | 'abPlane'
  | 'mandibularPlaneAngle'
  | 'yAxis'
  | 'cantOfOcclusion'
  | 'lowerIncisorToOcclusal'
  | 'impa'
  | 'interincisalAngle'
  | 'upperIncisalAngle';

export interface DownsStageValues {
  pre: number | '';
  mid: number | '';
  post: number | '';
}

export type DownsParametersMap = Record<DownsParameterKey, DownsStageValues>;

export interface DownsAnalysisData {
  parameters?: Partial<DownsParametersMap>;
  conclusion?: string;
}